remove mixin
